topdown/sets_bench_test: Move set builtin benchmarks.
